Check if .aliases and .zshrc exists to backup

+++ zsh/install.sh
@@ -13,9 +13,10 @@ echo "/> Configure zsh configurations"
 # Get the installation script directory path
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-cp ${HOME}/.zshrc ${HOME}/.zshrc_backup
+[[ -f ${HOME}/.zshrc ]] && cp ${HOME}/.zshrc ${HOME}/.zshrc_backup
 ln -sf ${DIR}/.zshrc ${HOME}
-cp ${HOME}/.aliases ${HOME}/.aliases_backup
+
+[[ -f ${HOME}/.aliases ]] && cp ${HOME}/.aliases ${HOME}/.aliases_backup
 ln -sf ${DIR}/.aliases ${HOME}
 
 # zsh as default shell
